separated the signup pages into two components
