+++ src/styles/EventSignUp.css
@@ -32,6 +32,7 @@ img {
     align-items: flex-start;
     padding: 0 3.5em;
     gap: 1em;
+    text-align: left;
 }
 
 h1 {
